+++ CHANGELOG.rst
@@ -7,6 +7,15 @@ smaht-portal
 Change Log
 ----------
 
+=======
+0.122.0
+=======
+`PR246: feat: homepage updates <https://github.com/smaht-dac/smaht-portal/pull/246>`_
+
+* Implement new announcements panel on homepage
+
+
+=======
 0.121.0
 =======
 `PR 300 SN Remove basecalling <https://github.com/smaht-dac/smaht-portal/pull/300>`